// src/app/models/DTO/calenderSheduale/MyCalenderEvent.ts
import {CalendarEvent} from "angular-calendar";
import {MyCalenderAction} from "./MyCalenderAction";

export class MyCalenderEvent  {


    id: number;
    title: string;
    titleEn: string;
    startHour: number;
    endHour: number;
    startDateLong: number;
    endDateLong: number;
    type: number;

}
 